+++ src/create_filtered_wav/create_filtered_wav.cpp
@@ -33,7 +33,7 @@ void ParseInputFile(FILE *fp, const SFormatChunk& formatChunk);
 void Destruct(const SFormatChunk& formatChunk);
 //void SetupWavHead(FILE* fp, const SFormatChunk& formatChunk, int dataSize);
 void Process(FILE *fp, FILE *theWavFile, const SFormatChunk& formatChunk, int dataSize);
-static void CopyFile(FILE* fpSrc, FILE* fpDst, long copySize);
+static void CopyFile(FILE* fpSrc, FILE* fpDst, int32_t copySize);
 
 int main(int argc, char* argv[])
 {
@@ -63,8 +63,8 @@ int main(int argc, char* argv[])
 	}
 
 	SFormatChunk formatChunkBuf;
-	long dataSize = ParseWaveHeader(fpWav, &formatChunkBuf);
-	long dataStartPos = ftell(fpWav);
+	int32_t dataSize = ParseWaveHeader(fpWav, &formatChunkBuf);
+	int32_t dataStartPos = ftell(fpWav);
 	rewind(fpWav);
 
 	if (dataSize > 0) {
@@ -271,7 +271,7 @@ void Process(FILE *fp, FILE *theWavFile, const SFormatChunk& formatChunk, int da
 enum { BUF_SIZE = 0x100000, };
 static unsigned char buf[BUF_SIZE];
 
-static void CopyFile(FILE* fpSrc, FILE* fpDst, long theSize)
+static void CopyFile(FILE* fpSrc, FILE* fpDst, int32_t theSize)
 {
 	size_t remainSize = theSize;
 	size_t copySize = BUF_SIZE;
